pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prism-scrnaseq"
version = "1.0.0"
description = "PRISM: Progenitor Resolution via Invariance-Sensitive Modeling"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Bryan Cheng"},
]
dependencies = [
    "torch>=2.0",
    "scanpy>=1.9",
    "anndata>=0.9",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.2",
    "numpyro>=0.13",
    "jax>=0.4",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "umap-learn>=0.5",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "einops>=0.6",
    "pandas>=1.5",
    "harmonypy>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
gpu = [
    "flash-attn>=2.0",
]

[project.scripts]
prism-run = "prism.cli:main"

[tool.setuptools.packages.find]
include = ["prism*"]

[tool.setuptools.package-data]
prism = ["../configs/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 120
target-version = "py39"
